Reorganise type definitions to easier documenting
