+++ .changeset/kind-apples-sniff.md
@@ -1,0 +1,5 @@
+---
+'@keystone-next/types': patch
+---
+
+Reorganised package internals to allow for easier documentation.
