+++ docs-next/CHANGELOG.md
@@ -1,19 +1,13 @@
 # @keystone-next/website
 
 ## 0.1.0
+
 ### Minor Changes
-
-
 
 - [`26543bd07`](https://github.com/keystonejs/keystone/commit/26543bd0752c470e336d61644c14e6a5333f65c0) [#4758](https://github.com/keystonejs/keystone/pull/4758) Thanks [@timleslie](https://github.com/timleslie)! - Added documentation for the `keystone-next` CLI.
 
-
 ### Patch Changes
 
-
-
 - [`08a67e820`](https://github.com/keystonejs/keystone/commit/08a67e820ff3f8ce22ba3fb57d1730202b18fddf) [#4764](https://github.com/keystonejs/keystone/pull/4764) Thanks [@timleslie](https://github.com/timleslie)! - Moved API docs into a separate section from the Guides.
-
-
 
 - [`4c635ae4b`](https://github.com/keystonejs/keystone/commit/4c635ae4bb5a4c4b5525103b55b145db7a443a63) [#4737](https://github.com/keystonejs/keystone/pull/4737) Thanks [@timleslie](https://github.com/timleslie)! - Added API placeholder.
