+++ packages-next/types/src/config/index.ts
@@ -1,0 +1,125 @@
+import type Knex from 'knex';
+import type { ConnectOptions } from 'mongoose';
+import { CorsOptions } from 'cors';
+import type { GraphQLSchema } from 'graphql';
+import { IncomingMessage } from 'http';
+
+import type { ListHooks } from './hooks';
+import type { ListAccessControl, FieldAccessControl } from './access-control';
+import type {
+  ListSchemaConfig,
+  ListConfig,
+  BaseFields,
+  FieldType,
+  FieldConfig,
+  MaybeSessionFunction,
+  MaybeItemFunction,
+  CacheHint,
+} from './lists';
+
+import type { KeystoneContext } from '..';
+import { CreateContext } from '../core';
+import type { BaseKeystone } from '../base';
+import { SessionStrategy } from '../session';
+import type { MaybePromise } from '../utils';
+
+export type KeystoneConfig = {
+  lists: ListSchemaConfig;
+  db: DatabaseConfig;
+  graphql?: GraphQLConfig;
+  session?: () => SessionStrategy<any>;
+  ui?: AdminUIConfig;
+  server?: ServerConfig;
+  extendGraphqlSchema?: ExtendGraphqlSchema;
+};
+
+// config.lists
+
+export type {
+  ListSchemaConfig,
+  ListConfig,
+  BaseFields,
+  FieldType,
+  FieldConfig,
+  MaybeSessionFunction,
+  MaybeItemFunction,
+  CacheHint,
+};
+
+// config.db
+
+export type DatabaseCommon = {
+  url: string;
+  onConnect?: (args: KeystoneContext) => Promise<void>;
+};
+
+export type DatabaseConfig = DatabaseCommon &
+  (
+    | {
+        adapter: 'prisma_postgresql';
+        dropDatabase?: boolean;
+        provider?: string;
+        getPrismaPath?: (arg: { prismaSchema: any }) => string;
+        getDbSchemaName?: (arg: { prismaSchema: any }) => string;
+        enableLogging?: boolean;
+      }
+    | {
+        adapter: 'knex';
+        dropDatabase?: boolean;
+        knexOptions?: { client?: string; connection?: string } & Knex.Config<any>;
+        schemaName?: string;
+      }
+    | { adapter: 'mongoose'; mongooseOptions?: { mongoUri?: string } & ConnectOptions }
+  );
+
+// config.graphql
+
+export type GraphQLConfig = {
+  // FIXME: We currently hardcode `/api/graphql` in a bunch of places
+  // We should be able to use config.graphql.path to set this path.
+  // path?: string;
+  queryLimits?: {
+    maxTotalResults?: number;
+  };
+};
+
+// config.server
+
+export type ServerConfig = {
+  /** Configuration options for the cors middleware. Set to `true` to use core Keystone defaults */
+  cors?: CorsOptions | true;
+  /** Port number to start the server on. Defaults to process.env.PORT || 3000 */
+  port?: number;
+};
+
+// config.ui
+
+export type AdminUIConfig = {
+  /** Enables certain functionality in the Admin UI that expects the session to be an item */
+  enableSessionItem?: boolean;
+  /** A function that can be run to validate that the current session should have access to the Admin UI */
+  isAccessAllowed?: (context: KeystoneContext) => MaybePromise<boolean>;
+  /** An array of page routes that can be accessed without passing the isAccessAllowed check */
+  publicPages?: string[];
+  /** The basePath for the Admin UI App */
+  path?: string;
+  getAdditionalFiles?: ((config: KeystoneConfig) => MaybePromise<AdminFileToWrite[]>)[];
+  pageMiddleware?: (args: {
+    req: IncomingMessage;
+    session: any;
+    isValidSession: boolean;
+    createContext: CreateContext;
+  }) => MaybePromise<{ kind: 'redirect'; to: string } | void>;
+};
+
+export type AdminFileToWrite =
+  | { mode: 'write'; src: string; outputPath: string }
+  | { mode: 'copy'; inputPath: string; outputPath: string };
+
+// config.extendGraphqlSchema
+
+export type ExtendGraphqlSchema = (schema: GraphQLSchema, keystone: BaseKeystone) => GraphQLSchema;
+
+// Exports from sibling packages
+
+export type { ListHooks, ListAccessControl, FieldAccessControl };
